+++ src/main/java/pl/kerpson/license/utilites/http/url/impl/LicensesURL.java
@@ -18,7 +18,7 @@ public class LicensesURL extends MURL {
   }
 
   @Override
-  public String getId(int id) {
+  public String getId(long id) {
     return String.format("https://api.mlicense.net/api/v1/licenses/id?id=%s", id);
   }
 
@@ -32,7 +32,7 @@ public class LicensesURL extends MURL {
   }
 
   @Override
-  public String delete(int id) {
+  public String delete(long id) {
     return String.format("https://api.mlicense.net/api/v1/licenses?id=%s", id);
   }
 }
